Change float64 to float32

+++ src/rlmcmc/env/_env.py
@@ -17,7 +17,7 @@ INF = 3.4028235e38  # Corresponds to the value of FLT_MAX in C++
 class RLMHEnvBase(gym.Env, ABC):
     def __init__(
         self,
-        log_target_pdf: Callable[[NDArray[np.float64]], np.float64],
+        log_target_pdf: Callable[[NDArray[np.float32]], np.float32],
         sample_dim: int = 2,
         total_timesteps: int = 10_000,
     ) -> None:
@@ -33,52 +33,52 @@ class RLMHEnvBase(gym.Env, ABC):
 
         # Observation specification
         self.observation_space = spaces.Box(
-            low=-INF, high=INF, shape=(2 * sample_dim,), dtype=np.float64
+            low=-INF, high=INF, shape=(2 * sample_dim,), dtype=np.float32
         )
 
         # Action specification
         self.action_space = spaces.Box(
-            low=-INF, high=INF, shape=(2 * sample_dim**2,), dtype=np.float64
+            low=-INF, high=INF, shape=(2 * sample_dim**2,), dtype=np.float32
         )
 
         # Store
-        self.store_observation: List[NDArray[np.float64]] = []
-        self.store_action: List[NDArray[np.float64]] = []
-        self.store_log_accetance_rate: List[np.float64] = []
+        self.store_observation: List[NDArray[np.float32]] = []
+        self.store_action: List[NDArray[np.float32]] = []
+        self.store_log_accetance_rate: List[np.float32] = []
         self.store_accetped_status: List[bool] = []
-        self.store_reward: List[np.float64] = []
+        self.store_reward: List[np.float32] = []
 
     def log_proposal_pdf(
         self,
-        x: NDArray[np.float64],
-        mean: NDArray[np.float64],
-        cov: NDArray[np.float64],
-    ) -> np.float64:
+        x: NDArray[np.float32],
+        mean: NDArray[np.float32],
+        cov: NDArray[np.float32],
+    ) -> np.float32:
         return multivariate_normal.logpdf(x, mean.flatten(), cov)
 
     @abstractmethod
     def distance_function(
-        self, current_sample: NDArray[np.float64], proposed_sample: NDArray[np.float64]
-    ) -> np.float64:
+        self, current_sample: NDArray[np.float32], proposed_sample: NDArray[np.float32]
+    ) -> np.float32:
         raise NotImplementedError("distance_function is not implemented.")
 
     @abstractmethod
     def reward_function(
         self,
-        current_sample: NDArray[np.float64],
-        proposed_sample: NDArray[np.float64],
-        log_alpha: np.float64,
-    ) -> np.float64:
+        current_sample: NDArray[np.float32],
+        proposed_sample: NDArray[np.float32],
+        log_alpha: np.float32,
+    ) -> np.float32:
         raise NotImplementedError("reward_function is not implemented.")
 
     def step(
-        self, action: NDArray[np.float64]
+        self, action: NDArray[np.float32]
     ) -> Tuple[
-        NDArray[np.float64],
-        np.float64,
+        NDArray[np.float32],
+        np.float32,
         bool,
         bool,
-        Dict[str, Union[NDArray[np.float64], bool, np.float64]],
+        Dict[str, Union[NDArray[np.float32], bool, np.float32]],
     ]:
         # Check Action Shape
         assert (
@@ -203,9 +203,9 @@ class RLMHEnvBase(gym.Env, ABC):
 
         # Store
         self.store_observation.append(self.observation)
-        self.store_log_accetance_rate.append(np.float64(0.0))
+        self.store_log_accetance_rate.append(np.float32(0.0))
         self.store_accetped_status.append(True)
-        self.store_reward.append(np.float64(0.0))
+        self.store_reward.append(np.float32(0.0))
 
         # Information Dictionary
         info = {
@@ -221,47 +221,47 @@ class RLMHEnvBase(gym.Env, ABC):
 class RLMHEnvV31(RLMHEnvBase):
     def __init__(
         self,
-        log_target_pdf: Callable[[NDArray[np.float64]], np.float64],
+        log_target_pdf: Callable[[NDArray[np.float32]], np.float32],
         sample_dim: int = 2,
         total_timesteps: int = 10_000,
     ) -> None:
         super().__init__(log_target_pdf, sample_dim, total_timesteps)
 
         # Add Additional Store
-        self._store_proposed_sample: List[NDArray[np.float64]] = []
-        self._store_current_sample: List[NDArray[np.float64]] = []
+        self._store_proposed_sample: List[NDArray[np.float32]] = []
+        self._store_current_sample: List[NDArray[np.float32]] = []
 
-        self._store_log_target_proposed: List[np.float64] = []
-        self._store_log_target_current: List[np.float64] = []
-        self._store_log_proposal_proposed: List[np.float64] = []
-        self._store_log_proposal_current: List[np.float64] = []
+        self._store_log_target_proposed: List[np.float32] = []
+        self._store_log_target_current: List[np.float32] = []
+        self._store_log_proposal_proposed: List[np.float32] = []
+        self._store_log_proposal_current: List[np.float32] = []
 
-        self._store_nearest_proposed_cov: List[NDArray[np.float64]] = []
-        self._store_nearest_current_cov: List[NDArray[np.float64]] = []
+        self._store_nearest_proposed_cov: List[NDArray[np.float32]] = []
+        self._store_nearest_current_cov: List[NDArray[np.float32]] = []
 
     def distance_function(
-        self, current_sample: NDArray[np.float64], proposed_sample: NDArray[np.float64]
-    ) -> np.float64:
+        self, current_sample: NDArray[np.float32], proposed_sample: NDArray[np.float32]
+    ) -> np.float32:
         return np.linalg.norm(current_sample - proposed_sample, 2)
 
     def reward_function(
         self,
-        current_sample: NDArray[np.float64],
-        proposed_sample: NDArray[np.float64],
-        log_alpha: np.float64,
-    ) -> np.float64:
+        current_sample: NDArray[np.float32],
+        proposed_sample: NDArray[np.float32],
+        log_alpha: np.float32,
+    ) -> np.float32:
         return np.power(
             self.distance_function(current_sample, proposed_sample), 2
         ) * np.exp(log_alpha)
 
     def step(
-        self, action: NDArray[np.float64]
+        self, action: NDArray[np.float32]
     ) -> Tuple[
-        NDArray[np.float64],
-        np.float64,
+        NDArray[np.float32],
+        np.float32,
         bool,
         bool,
-        Dict[str, Union[NDArray[np.float64], bool, np.float64]],
+        Dict[str, Union[NDArray[np.float32], bool, np.float32]],
     ]:
         # Check Action Shape
         assert (
@@ -379,20 +379,20 @@ class RLMHEnvV31(RLMHEnvBase):
 
 class RLMHEnvV31A(RLMHEnvV31):
     def distance_function(
-        self, current_sample: NDArray[np.float64], proposed_sample: NDArray[np.float64]
-    ) -> np.float64:
+        self, current_sample: NDArray[np.float32], proposed_sample: NDArray[np.float32]
+    ) -> np.float32:
         return np.min(np.abs(current_sample - proposed_sample))
 
 
 class RLMHEnvV31B(RLMHEnvV31A):
     def reward_function(
         self,
-        current_sample: NDArray[np.float64],
-        proposed_sample: NDArray[np.float64],
-        log_alpha: np.float64,
+        current_sample: NDArray[np.float32],
+        proposed_sample: NDArray[np.float32],
+        log_alpha: np.float32,
         weight: float = 0.7,
         kappa: float = 1.0,
-    ) -> np.float64:
+    ) -> np.float32:
         return weight * expit(
             kappa
             * self.distance_function(current_sample, proposed_sample)
